Refactor Tool and Workload command subcommands - separate parsing and action logic
