+++ src/Cli/dotnet/Commands/Tool/Run/ToolRunCommandDefinition.cs
@@ -1,0 +1,39 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+#nullable disable
+
+using System.CommandLine;
+
+namespace Microsoft.DotNet.Cli.Commands.Tool.Run;
+
+internal static class ToolRunCommandDefinition
+{
+    public static readonly Argument<string> CommandNameArgument = new("commandName")
+    {
+        HelpName = CliCommandStrings.CommandNameArgumentName,
+        Description = CliCommandStrings.CommandNameArgumentDescription
+    };
+
+    public static readonly Argument<IEnumerable<string>> CommandArgument = new("toolArguments")
+    {
+        Description = CliCommandStrings.ToolRunArgumentsDescription
+    };
+
+    public static readonly Option<bool> RollForwardOption = new("--allow-roll-forward")
+    {
+        Description = CliCommandStrings.RollForwardOptionDescription,
+        Arity = ArgumentArity.Zero
+    };
+
+    public static Command Create()
+    {
+        Command command = new("run", CliCommandStrings.ToolRunCommandDescription);
+
+        command.Arguments.Add(CommandNameArgument);
+        command.Arguments.Add(CommandArgument);
+        command.Options.Add(RollForwardOption);
+
+        return command;
+    }
+}
