+++ src/Cli/dotnet/Commands/Tool/Search/ToolSearchCommandParser.cs
@@ -9,35 +9,15 @@ namespace Microsoft.DotNet.Cli.Commands.Tool.Search;
 
 internal static class ToolSearchCommandParser
 {
-    public static readonly Argument<string> SearchTermArgument = new("searchTerm")
-    {
-        HelpName = CliCommandStrings.ToolSearchSearchTermArgumentName,
-        Description = CliCommandStrings.ToolSearchSearchTermDescription
-    };
+    public static readonly Argument<string> SearchTermArgument = ToolSearchCommandDefinition.SearchTermArgument;
 
-    public static readonly Option<bool> DetailOption = new("--detail")
-    {
-        Description = CliCommandStrings.DetailDescription,
-        Arity = ArgumentArity.Zero
-    };
+    public static readonly Option<bool> DetailOption = ToolSearchCommandDefinition.DetailOption;
 
-    public static readonly Option<string> SkipOption = new("--skip")
-    {
-        Description = CliCommandStrings.ToolSearchSkipDescription,
-        HelpName = CliCommandStrings.ToolSearchSkipArgumentName
-    };
+    public static readonly Option<string> SkipOption = ToolSearchCommandDefinition.SkipOption;
 
-    public static readonly Option<string> TakeOption = new("--take")
-    {
-        Description = CliCommandStrings.ToolSearchTakeDescription,
-        HelpName = CliCommandStrings.ToolSearchTakeArgumentName
-    };
+    public static readonly Option<string> TakeOption = ToolSearchCommandDefinition.TakeOption;
 
-    public static readonly Option<bool> PrereleaseOption = new("--prerelease")
-    {
-        Description = CliCommandStrings.ToolSearchPrereleaseDescription,
-        Arity = ArgumentArity.Zero
-    };
+    public static readonly Option<bool> PrereleaseOption = ToolSearchCommandDefinition.PrereleaseOption;
 
     private static readonly Command Command = ConstructCommand();
 
@@ -48,14 +28,7 @@ internal static class ToolSearchCommandParser
 
     private static Command ConstructCommand()
     {
-        Command command = new("search", CliCommandStrings.ToolSearchCommandDescription);
-
-        command.Arguments.Add(SearchTermArgument);
-
-        command.Options.Add(DetailOption);
-        command.Options.Add(SkipOption);
-        command.Options.Add(TakeOption);
-        command.Options.Add(PrereleaseOption);
+        Command command = ToolSearchCommandDefinition.Create();
 
         command.SetAction((parseResult) => new ToolSearchCommand(parseResult).Execute());
