+++ src/Cli/dotnet/Commands/Workload/Search/WorkloadSearchCommandDefinition.cs
@@ -1,0 +1,31 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+#nullable disable
+
+using System.CommandLine;
+
+namespace Microsoft.DotNet.Cli.Commands.Workload.Search;
+
+internal static class WorkloadSearchCommandDefinition
+{
+    public static readonly Argument<string> WorkloadIdStubArgument =
+        new(CliCommandStrings.WorkloadIdStubArgumentName)
+        {
+            Arity = ArgumentArity.ZeroOrOne,
+            Description = CliCommandStrings.WorkloadIdStubArgumentDescription
+        };
+
+    public static readonly Option<string> VersionOption = InstallingWorkloadCommandParser.VersionOption;
+
+    public static Command Create()
+    {
+        var command = new Command("search", CliCommandStrings.WorkloadSearchCommandDescription);
+        command.Subcommands.Add(WorkloadSearchVersionsCommandParser.GetCommand());
+        command.Arguments.Add(WorkloadIdStubArgument);
+        command.Options.Add(CommonOptions.HiddenVerbosityOption);
+        command.Options.Add(VersionOption);
+
+        return command;
+    }
+}
